Add more docs to TargetHandler
